feat: export a list of modules for use in @storybook/react-native 6 (#122)
* feat: export a list of modules for use in @storybook/react-native 6

In @storybook/react-native version 6: the loader function passed to 'configure' should return void or an array of module exports that all contain a 'default' export. `loadStories()` just loads all the stories and `stories` is just a list of the story locations.  Adding `modules` so that we can also access a list of module exports.

* typo

* add modules export to readme

* feat: add modules to loader

* feat: use modules instead of storyFiles

* update storyFiles

* update snapshot

* fix: modules value on snapshot

* fix: return to storyFiles

* fix: remove file extensions

* fix: remove file extensions

* fix: trailing comma